chore(core): move values helper to data module
The values() helper started as a little lookup utility, but now it's doing all sorts of other stuff – running transforms, resolving multiple datasets, resolving raw primitives, and so on. It makes more sense for this to live in the data module.
