+++ source/scales.js
@@ -1,9 +1,9 @@
 import * as d3 from 'd3'
-import { data, sumByCovariates } from './data.js'
+import { data, sumByCovariates, values } from './data.js'
 import { colors } from './color.js'
 import { encodingChannelQuantitative, encodingType, encodingValue } from './encodings.js'
 import { feature } from './feature.js'
-import { identity, isDiscrete, isTextChannel, values } from './helpers.js'
+import { identity, isDiscrete, isTextChannel } from './helpers.js'
 import { memoize } from './memoize.js'
 import { parseTime, temporalBarDimensions } from './time.js'
 import { sorter } from './sort.js'
